backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
python-multipart
python-dotenv
requests
pandas
paho-mqtt
redis>=4.0.0